requirements.txt
langchain
langchain_openai

python-dotenv
langchain_community

langchain_core

streamlit

langchain_groq
langchain-huggingface
faiss-cpu
sentence-transformers